requirements.txt
# Story Forge - Self-Publishing Dashboard
# Python dependencies

# Core
fastapi>=0.100.0
uvicorn>=0.20.0

# Database
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0

# Security
keyring>=25.0.0
cryptography>=46.0.0

# Authentication
authlib>=1.0.0
google-auth>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0.0

# TTS Providers
httpx>=0.27.0
elevenlabs>=1.0.0

# Manuscript Export
python-docx>=1.1.0

# Dev / Testing
playwright>=1.40.0
pytest>=8.0.0
